Graphs included for Quarter chord and mid chord angles optimization
An important step towards the Aerodynamics code. The Cl_alpha has been determined, while the downwash angle per angle of attack is yet to be determine. The decision will be based on a SolidWorks Top View model of the sizing that has be concluded by the code. Furthermore, RC Model manufacturing must be studied prior to make a decision on the sizing.